polycubed: fix list validation
The validation code for the list was not testing all the elements of the list,
it instead was testing the list as a container.  Due to this complex commands
issued with a rest client were failing, for instance:

curl -L -X POST --data '{"name":"hw0","ports":[{"name":"pippo"}]}' \
  "http://localhost:9000/polycube/v1/helloworld/hw0/"

curl -L -X POST --data \
  '[{"name":"toveth1", "peer":"veth1"}, {"name":"toveth2", "peer":"veth2"}]' \
  "http://localhost:9000/polycube/v1/helloworld/hw0/ports/

This commit implements a new logic that tests all the elements in a list, also
a new logic fill the keys in json body of a list request is added as well.

This failure was unnoticed because polycubectl does not allow to perform such
complex operations in a single command.

Fixes: d8d280cd96a4 ("Added new REST API and validation code")
Fixes: 795ce964c60c ("move help generation from services to daemon")

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>